[FIX] gitignore tracking DS_Store file.
the problem you're having is that in some earlier commit, you've accidentally added .DS_Store files to the repository. Of course, once a file is tracked in your repository, it will continue to be tracked even if it matches an entry in an applicable .gitignore file.

This commit manually removes the .DS_Store files that was added to the repository.